desktop/requirements.txt
customtkinter
opencv-python
mediapipe==0.10.9
torch
torchvision
torchaudio
pillow
numpy
